refactor: make error-check-macro reusable in the package
