+++ include/cuda_blackboard/cuda_unique_ptr.hpp
@@ -20,30 +20,18 @@
 #ifndef CUDA_BLACKBOARD__CUDA_UNIQUE_PTR_HPP_
 #define CUDA_BLACKBOARD__CUDA_UNIQUE_PTR_HPP_
 
+#include "cuda_blackboard/cuda_error.hpp"
+
 #include <cuda_runtime.h>
 #include <cuda_runtime_api.h>
 
+#include <cstddef>
+#include <functional>
 #include <memory>
-#include <sstream>
-#include <stdexcept>
 #include <type_traits>
-
-#define CUDA_BLACKBOARD_CHECK_CUDA_ERROR(e) \
-  (cuda_blackboard::cuda_check_error(e, __FILE__, __LINE__))
 
 namespace cuda_blackboard
 {
-
-template <typename F, typename N>
-void cuda_check_error(const ::cudaError_t e, F && f, N && n)
-{
-  if (e != ::cudaSuccess) {
-    std::stringstream s;
-    s << ::cudaGetErrorName(e) << " (" << e << ")@" << f << "#L" << n << ": "
-      << ::cudaGetErrorString(e);
-    throw std::runtime_error{s.str()};
-  }
-}
 
 struct CudaDeleter
 {
